auth-service/requirements.txt
Django>=4.2.7,<4.3
djangorestframework>=3.14.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
